test: consolidate the type-audit round's scattered test files
The type-coverage audit round (#62-#69) added many small single-purpose
test files. Fold them into thematic homes without changing any test
logic (verified: 210 -> 210 test funcs, full suite green). Net -11 files
(35 -> 24).

Black-box (convert_test):
- identity_key      -> hashable_test (map keys)
- decorate_key      -> order_test (key ordering)
- func_pointer      -> panics_test (panic prevention)
- gointerface_nil   -> interface_test (GoInterface)
- collection_elem,
  uintptr           -> checked_conv_test (checked conversion)
- typed_nil,
  low_misc          -> review_misc_test (misc)

White-box (convert):
- stable_key,
  truth_nilable     -> internal_test

Root (starlight):
- eval_nil_reader   -> starlight_test
