+++ convert/internal_test.go
@@ -151,3 +151,67 @@ func TestComparableByValue(t *testing.T) {
 		}
 	}
 }
+
+// TestStableKeyString covers every kind branch of the address-free key
+// renderer and asserts equal values render equally (no addresses leak).
+func TestStableKeyString(t *testing.T) {
+	type inner struct {
+		B bool
+		F float64
+	}
+	type composite struct {
+		N int
+		U uint8
+		S string
+		P *int
+		C complex128
+		A [2]int
+		I inner
+	}
+	mk := func() composite {
+		x := 7
+		return composite{N: -3, U: 200, S: "k", P: &x, C: 1 + 2i, A: [2]int{4, 5}, I: inner{B: true, F: 1.5}}
+	}
+	a, b := mk(), mk() // equal values, different *int addresses
+	sa := stableKeyString(reflect.ValueOf(a))
+	sb := stableKeyString(reflect.ValueOf(b))
+	if sa != sb {
+		t.Fatalf("equal values must render equally:\n a=%s\n b=%s", sa, sb)
+	}
+	for _, want := range []string{"-3", "200", "k", "7", "true", "1.5", "[4 5]"} {
+		if !strings.Contains(sa, want) {
+			t.Errorf("rendered key %q missing %q", sa, want)
+		}
+	}
+	// nil pointer and channel branches
+	if got := stableKeyString(reflect.ValueOf((*int)(nil))); got != "<nil>" {
+		t.Errorf("nil ptr -> %q, want <nil>", got)
+	}
+	if got := stableKeyString(reflect.ValueOf(make(chan int))); got != "<chan>" {
+		t.Errorf("chan -> %q, want <chan>", got)
+	}
+}
+
+// TestGoInterfaceTruthNilable: Truth returned a blanket true for kinds it
+// didn't enumerate, so a nil chan/map/func/slice wrapped in a GoInterface
+// reported truthy. Nilable kinds should reflect their nil-ness.
+func TestGoInterfaceTruthNilable(t *testing.T) {
+	cases := []struct {
+		v    interface{}
+		want bool
+	}{
+		{(chan int)(nil), false},
+		{make(chan int), true},
+		{map[string]int(nil), false},
+		{map[string]int{"a": 1}, true},
+		{([]int)(nil), false},
+		{[]int{1}, true},
+		{(func())(nil), false},
+	}
+	for _, c := range cases {
+		gi := &GoInterface{v: reflect.ValueOf(c.v)}
+		if got := bool(gi.Truth()); got != c.want {
+			t.Errorf("Truth(%T nil=%v) = %v, want %v", c.v, reflect.ValueOf(c.v).IsNil(), got, c.want)
+		}
+	}
+}
